yarn: Add "rollback:sql" to undo database changes
This commit also implements the feedback given in
https://github.com/serlo/serlo.org/pull/129/files#r362140621
