+++ scripts/transform/index.ts
@@ -1,0 +1,2 @@
+export { ConcatenateInsertCommands } from './ConcatenateInsertCommands'
+export { IgnoreInsecurePasswordWarning } from './IgnoreInsecurePasswordWarning'
